[AUTO-21] add `StudentsLayout`
