+++ src/app/[locale]/dash/admin/(users)/students/page.tsx
@@ -1,8 +1,5 @@
-import { getTranslations } from 'next-intl/server';
-
 import { getStudentFolder } from '@/base/utils/server/students/get-student-folder';
 import { StudentFile } from '@/components/sections/students';
-import type { Locale } from '@/base/data/locales';
 
 import StudentsPage from './_components/students';
 import StudentNotFound from './_components/not-found';
@@ -25,21 +22,4 @@ export default async function Students({
   if (!student) return <StudentNotFound />;
 
   return <StudentFile student={student} />;
-}
-
-export async function generateMetadata({
-  params: { locale },
-}: {
-  params: {
-    locale: Locale;
-  };
-}) {
-  const t = await getTranslations({
-    locale,
-    namespace: 'Dashboard.Users.Students.Header',
-  });
-
-  return {
-    title: `${t('title')} / Dashboard`,
-  };
 }
